# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BBPW Football Prediction Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  /** @pattern ^[0-9]{4}$ */
  pin: string;
}

export interface LoginInput {
  username: string;
  pin: string;
  rememberMe?: boolean;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  coins: number;
  wins: number;
  losses: number;
  accuracy: number;
  /** @nullable */
  rank?: number | null;
  isAdmin: boolean;
  createdAt: string;
  /** @nullable */
  lastLoginAt?: string | null;
  /** @nullable */
  lastDailyRewardAt?: string | null;
  badges: string[];
}

export interface AuthResponse {
  user: User;
  token: string;
}

/**
 * @nullable
 */
export type UserProfileFriendStatus = typeof UserProfileFriendStatus[keyof typeof UserProfileFriendStatus] | null;


export const UserProfileFriendStatus = {
  none: 'none',
  friends: 'friends',
  pending_sent: 'pending_sent',
  pending_received: 'pending_received',
  null: 'null',
} as const;

export type PredictionPrediction = typeof PredictionPrediction[keyof typeof PredictionPrediction];


export const PredictionPrediction = {
  home: 'home',
  draw: 'draw',
  away: 'away',
} as const;

/**
 * @nullable
 */
export type PredictionOutcome = typeof PredictionOutcome[keyof typeof PredictionOutcome] | null;


export const PredictionOutcome = {
  win: 'win',
  loss: 'loss',
  pending: 'pending',
  null: 'null',
} as const;

export interface Prediction {
  id: number;
  userId: number;
  matchId: number;
  prediction: PredictionPrediction;
  stake: number;
  /** @nullable */
  payout?: number | null;
  /** @nullable */
  outcome?: PredictionOutcome;
  createdAt: string;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  upcoming: 'upcoming',
  live: 'live',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

/**
 * @nullable
 */
export type MatchResult = typeof MatchResult[keyof typeof MatchResult] | null;


export const MatchResult = {
  home: 'home',
  draw: 'draw',
  away: 'away',
  null: 'null',
} as const;

export interface Match {
  id: number;
  league: string;
  leagueLabel: string;
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  kickoffAt: string;
  status: MatchStatus;
  /** @nullable */
  result?: MatchResult;
  totalPool: number;
  predictionCount: number;
  createdAt: string;
}

export interface PredictionWithMatch {
  prediction: Prediction;
  match: Match;
}

export interface UserProfile {
  user: User;
  recentPredictions: PredictionWithMatch[];
  /** @nullable */
  leaderboardRank?: number | null;
  /** @nullable */
  friendStatus?: UserProfileFriendStatus;
}

export interface UserUpdate {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username?: string;
  /** @nullable */
  avatarUrl?: string | null;
  isAdmin?: boolean;
  coins?: number;
}

export interface MatchDetail {
  match: Match;
  userPrediction?: Prediction;
  homeVotes: number;
  drawVotes: number;
  awayVotes: number;
  totalStaked: number;
}

export interface MatchInput {
  league: string;
  leagueLabel: string;
  homeTeam: string;
  awayTeam: string;
  kickoffAt: string;
}

export type MatchUpdateStatus = typeof MatchUpdateStatus[keyof typeof MatchUpdateStatus];


export const MatchUpdateStatus = {
  upcoming: 'upcoming',
  live: 'live',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface MatchUpdate {
  league?: string;
  leagueLabel?: string;
  homeTeam?: string;
  awayTeam?: string;
  kickoffAt?: string;
  status?: MatchUpdateStatus;
}

export type ResultInputResult = typeof ResultInputResult[keyof typeof ResultInputResult];


export const ResultInputResult = {
  home: 'home',
  draw: 'draw',
  away: 'away',
} as const;

export interface ResultInput {
  result: ResultInputResult;
  homeScore: number;
  awayScore: number;
}

export interface SettlementResult {
  match: Match;
  winnersCount: number;
  losersCount: number;
  totalPool: number;
  payoutPerWinner: number;
}

export type PredictionInputPrediction = typeof PredictionInputPrediction[keyof typeof PredictionInputPrediction];


export const PredictionInputPrediction = {
  home: 'home',
  draw: 'draw',
  away: 'away',
} as const;

export interface PredictionInput {
  matchId: number;
  prediction: PredictionInputPrediction;
  /** @minimum 10 */
  stake: number;
}

export interface LeaderboardEntry {
  rank: number;
  user: User;
  coins: number;
  wins: number;
  accuracy: number;
}

export interface Friend {
  id: number;
  user: User;
  friendedAt: string;
}

export type FriendRequestStatus = typeof FriendRequestStatus[keyof typeof FriendRequestStatus];


export const FriendRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface FriendRequest {
  id: number;
  fromUser: User;
  toUser: User;
  status: FriendRequestStatus;
  createdAt: string;
}

export interface FriendRequestInput {
  toUserId: number;
}

export type FriendRequestResponseAction = typeof FriendRequestResponseAction[keyof typeof FriendRequestResponseAction];


export const FriendRequestResponseAction = {
  accept: 'accept',
  decline: 'decline',
} as const;

export interface FriendRequestResponse {
  action: FriendRequestResponseAction;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  prediction_won: 'prediction_won',
  prediction_lost: 'prediction_lost',
  friend_request: 'friend_request',
  match_reminder: 'match_reminder',
  daily_reward: 'daily_reward',
  system: 'system',
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface DailyRewardResult {
  coinsAwarded: number;
  newBalance: number;
  streak: number;
}

export interface ActivityItem {
  type: string;
  description: string;
  timestamp: string;
}

export interface AdminStats {
  totalUsers: number;
  totalMatches: number;
  activeMatches: number;
  totalPredictions: number;
  totalCoinsInCirculation: number;
  recentActivity: ActivityItem[];
}

export interface CoinUpdate {
  /** @minimum 0 */
  coins: number;
}

export type SearchUsersParams = {
q: string;
};

export type ListMatchesParams = {
status?: ListMatchesStatus;
};

export type ListMatchesStatus = typeof ListMatchesStatus[keyof typeof ListMatchesStatus];


export const ListMatchesStatus = {
  upcoming: 'upcoming',
  live: 'live',
  completed: 'completed',
  all: 'all',
} as const;

export type GetLeaderboardParams = {
sortBy?: GetLeaderboardSortBy;
limit?: number;
};

export type GetLeaderboardSortBy = typeof GetLeaderboardSortBy[keyof typeof GetLeaderboardSortBy];


export const GetLeaderboardSortBy = {
  coins: 'coins',
  accuracy: 'accuracy',
  wins: 'wins',
} as const;