# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BBPW Football Prediction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user (username + 4-digit PIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and 4-digit PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /friends/requests:
    get:
      operationId: getFriendRequests
      tags: [friends]
      summary: Get pending friend requests
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
  /friends/requests/{requestId}:
    patch:
      operationId: respondFriendRequest
      tags: [friends]
      summary: Accept or decline a friend request
      parameters:
        - in: path
          name: requestId
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestResponse"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/{friendId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - in: path
          name: friendId
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/users/{userId}/coins:
    patch:
      operationId: setUserCoins
      tags: [admin]
      summary: Set a user's coin balance (admin only)
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoinUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{userId}/toggle-admin:
    patch:
      operationId: toggleAdmin
      tags: [admin]
      summary: Toggle admin status for a user (admin only)
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        pin:
          type: string
          pattern: "^[0-9]{4}$"
      required:
        - username
        - pin
    LoginInput:
      type: object
      properties:
        username:
          type: string
        pin:
          type: string
        rememberMe:
          type: boolean
      required:
        - username
        - pin
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        coins:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        accuracy:
          type: number
        rank:
          type: ["integer", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
        lastDailyRewardAt:
          type: ["string", "null"]
          format: date-time
        badges:
          type: array
          items:
            type: string
      required:
        - id
        - username
        - coins
        - wins
        - losses
        - accuracy
        - isAdmin
        - createdAt
        - badges
...
    FriendRequest:
      type: object
      properties:
        id:
          type: integer
        fromUser:
          $ref: "#/components/schemas/User"
        toUser:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fromUser
        - toUser
        - status
        - createdAt
    FriendRequestInput:
      type: object
      properties:
        toUserId:
          type: integer
      required:
        - toUserId
    FriendRequestResponse:
      type: object
      properties:
        action:
          type: string
          enum: [accept, decline]
      required:
        - action
...
    CoinUpdate:
      type: object
      properties:
        coins:
          type: integer
          minimum: 0
      required:
        - coins