# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BBPW Football Prediction Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  coins: number;
  wins: number;
  losses: number;
  accuracy: number;
  /** @nullable */
  rank?: number | null;
  isAdmin: boolean;
  createdAt: Date;
  /** @nullable */
  lastLoginAt?: Date | null;
  /** @nullable */
  lastDailyRewardAt?: Date | null;
  badges: string[];
}